Moved Offline and OnlineServiceProvider to DAL namespace
