+++ MoodMovies/DataAccessLayer/OfflineServiceProvider.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MoodMovies.Logic
+namespace MoodMovies.DataAccessLayer
 {
     public class OfflineServiceProvider : IServiceProvider, IOfflineServiceProvider
     {
