+++ MoodMovies/MoodMovies.csproj
@@ -130,11 +130,11 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Logic\ImageCacher.cs" />
-    <Compile Include="Logic\IOfflineServiceProvider.cs" />
-    <Compile Include="Logic\IOnlineServiceProvider.cs" />
-    <Compile Include="Logic\IServiceProvider.cs" />
-    <Compile Include="Logic\OfflineServiceProvider.cs" />
-    <Compile Include="Logic\OnlineServiceProvider.cs" />
+    <Compile Include="DataAccessLayer\IOfflineServiceProvider.cs" />
+    <Compile Include="DataAccessLayer\IOnlineServiceProvider.cs" />
+    <Compile Include="DataAccessLayer\IServiceProvider.cs" />
+    <Compile Include="DataAccessLayer\OfflineServiceProvider.cs" />
+    <Compile Include="DataAccessLayer\OnlineServiceProvider.cs" />
     <Compile Include="Messages\AccountMessages.cs" />
     <Compile Include="Messages\ClientChangeMessage.cs" />
     <Compile Include="Messages\MovieCardMessages.cs" />
