+++ MoodMovies/Services/SearchService.cs
@@ -1,5 +1,5 @@
 ﻿using DataModel.DataModel.Entities;
-using MoodMovies.Logic;
+using MoodMovies.DataAccessLayer;
 using System;
 using System.Collections.Generic;
 using System.Linq;
